+++ src/core/use-cases/list-todo/list-todo.ts
@@ -1,13 +1,13 @@
 import { Todo } from '../../entities';
 import { ITodoRepository } from '../../ports';
 import { IListTodoUseCase } from '..';
-import { UseCaseContext } from '../../../main/container';
+import { InfraContext } from '../../../main/container';
 
 export class ListTodoUseCase implements IListTodoUseCase {
   private todo_repository: ITodoRepository;
 
-  constructor(use_case_context: UseCaseContext) {
-    this.todo_repository = use_case_context.todo_repository;
+  constructor(infra_context: InfraContext) {
+    this.todo_repository = infra_context.todo_repository;
   }
 
   list(): Todo[] {