+++ src/interface/amqp/consumers/create-todo/create-todo.dto.ts
@@ -1,0 +1,7 @@
+import { TodoStatus } from '../../../../core/entities';
+
+export type TodoMessage = {
+  name: string;
+  status: TodoStatus;
+  user: string;
+};