+++ src/main/app.spec.ts
@@ -12,13 +12,18 @@ describe('App', () => {
       const http_mock = {
         start: sinon.fake.returns(undefined),
       };
+
+      const amqp_mock = {
+        start: sinon.fake.returns(undefined),
+      };
       // @ts-ignore
-      const app = new App({ cli: cli_mock, http: http_mock });
+      const app = new App({ cli: cli_mock, http: http_mock, amqp: amqp_mock });
 
       app.start();
 
       assert(cli_mock.start.calledOnce);
       assert(http_mock.start.calledOnce);
+      assert(amqp_mock.start.calledOnce);
     });
   });
 });