+++ src/main/modules/index.ts
@@ -1,3 +1,4 @@
 export * from './modules.interface';
 export * from './cli/cli';
 export * from './http/http';
+export * from './amqp/amqp';