+++ src/shared/index.ts
@@ -1,1 +1,3 @@
-export * from './coded-error';
+export * from './errors/coded-error';
+export * from './errors/validation-error';
+export * from './helper/bufferConverter/buffer-converter';